Routing and map.onLoad: remove unneeded global-ness:
No need for dispatch if leaning entirely on routing (aka filtering the
stored features for the matched ID). Maybe this is bad practice but
there is no noticeable performance hit and the Details panel is
literally the only place the feature details are needed, so it makes
sense to keep filtering in that component.