Routing: map clickout to deselect: WHY DOES THIS WORK:
This might be one of the strangest and most confusing pecularities I
have ever encountered as a programmer. There is no discernible reason
why this works, but it does. Future Jason must trim it down and
understand, otherwise there can be no order or balance in the universe.

But at least it works...