Routing: about as good as we're going to get for now:
Lost the "flyTo" on load when feature is selected, but the main things
(back/fwd, on load, click, etc.) are working properly. The only big
annoying this is that the de-selection is still part of the routing, so
it's silly if back/fwd are used. Might be edge case though, have to just
keep momentum for now and hope that solution appears later.