More refactoring, mainly census types cleanup, Sheets-to-JSON improvements:
The mixture of Sheets API response types and those combined with
react-query types and/or very specific types (e.g. just a two-column
situation) was confusing and misleading so renamed a bunch of types.
This rabbit hole started with the need to make Sheets column order
irrelevant and instead rely on heading names from the Sheets-to-JSON
conversion, which seems to resemble a real-world API or CMS more
closely.

The only thing that should now rely on column order should be the
MB_Fonts sheet, but that shouldn't be around long as the switch is made
to Language Config and its abundant headings.