Big fella: legend: nearly all config from Airtable; improve grouped legend:
For that, use just a single component and rely on a "groupName" and an
array of items to populate it.

Legend config is now just a handful of lines, and the majority is coming
from Airtable. The map symbology (and lots of global-ish state) are
still using the original setup, but we're getting there.