/Explore/Neighborhood: correctly show Additional neighborhoods:
Airtable's FIND formula returns things like "East Elmhurst" in an
"Elmhurst" query so, in lieu of a good formula hack we are now
processing it on the client side via an `addlNeighborhoods` filter.

This commit also attempts to tighten up, or at least decouple, the
"field" vs. "tableName" vs. "routeable table names". Really only the
last one is needed as a type in most or all cases.