Theme: use new color palette, standardize link color (#189, #190):
The primary palette is more or less the same, not sure if it was just a
discrepancy with Color Picker on the designer's style guide or
intentional on her part.

The secondary palette is new and has been applied to links and other
components that may have not been secondary before.