Restore point tooltips and polygon highlighting on hover:
Some major caveats to this, like can't prevent highlight on mobile, and
the route-based approach to preventing clearing of selected feature is
really fragile, and hovering outside of a polygon (e.g. on basemap) does
not clear any previously highlighted feature.

Other than that, pretty good, as long as devices don't start on fire
from all that's going on...